Conflicts:
	sugarcrm/clients/base/fields/enum/enum.js

+++ sugarcrm/tests/service/Bug51617Test.php
@@ -36,6 +36,7 @@ class Bug51617Test extends SOAPTestCase
         SugarTestHelper::setUp('beanFiles');
         SugarTestHelper::setUp('beanList');
         SugarTestHelper::setUp('current_user');
+
         $GLOBALS['current_user']->status = 'Active';
         $GLOBALS['current_user']->is_admin = 1;
         $GLOBALS['current_user']->save();
@@ -43,7 +44,7 @@ class Bug51617Test extends SOAPTestCase
         $this->field = get_widget('varchar');
         $this->field->id = 'Accountstest_custom_c';
         $this->field->name = 'test_custom_c';
-        $this->field->vanme = 'LBL_TEST_CUSTOM_C';
+        $this->field->vname = 'LBL_TEST_CUSTOM_C';
         $this->field->comments = NULL;
         $this->field->help = NULL;
         $this->field->custom_module = 'Accounts';
@@ -69,6 +70,7 @@ class Bug51617Test extends SOAPTestCase
         $this->df->setup($this->mod);
         $this->df->addFieldObject($this->field);
         $this->df->buildCache('Accounts');
+        $GLOBALS['db']->commit();
         VardefManager::clearVardef();
         VardefManager::refreshVardefs('Accounts', 'Account');
         $this->mod->field_defs = $GLOBALS['dictionary']['Account']['fields'];
@@ -76,6 +78,8 @@ class Bug51617Test extends SOAPTestCase
         $this->_account = SugarTestAccountUtilities::createAccount();
 
         $this->_account->test_custom_c = 'Custom Field';
+        $this->_account->team_set_id = '1';
+        $this->_account->team_id = '1';
         $this->_account->save();
 
         $GLOBALS['db']->commit(); // Making sure we commit any changes
@@ -86,13 +90,17 @@ class Bug51617Test extends SOAPTestCase
     public function tearDown()
     {
         $this->df->deleteField($this->field);
-        $GLOBALS['db']->query("DELETE FROM accounts_cstm WHERE id_c = '{$this->_account->id}'");
+        if ($GLOBALS['db']->tableExists('accounts_cstm')) {
+            $GLOBALS['db']->query("DELETE FROM accounts_cstm WHERE id_c = '{$this->_account->id}'");
+        }
 
         SugarTestAccountUtilities::removeAllCreatedAccounts();
         SugarTestUserUtilities::removeAllCreatedAnonymousUsers();
         SugarTestHelper::tearDown();
+        unset($GLOBALS['current_user']);
 
         parent::tearDown();
+        $GLOBALS['db']->commit();
 
         global $soap_version_test_accountId, $soap_version_test_opportunityId, $soap_version_test_contactId;
         unset($soap_version_test_accountId);
@@ -106,11 +114,12 @@ class Bug51617Test extends SOAPTestCase
     public function testGetEntryListWithCustomField()
     {
         $this->_login();
+        $GLOBALS['db']->commit();
         $result = $this->_soapClient->call('get_entry_list',
             array(
                  'session'=>$this->_sessionId,
                  "module_name" => 'Accounts',
-                 '',
+                 "id='".$this->_account->id."'",
                  '',
                  0,
                  "select_fields" => array('id', 'name', 'test_custom_c'),
@@ -130,10 +139,10 @@ class Bug51617Test extends SOAPTestCase
         {
             foreach($row as $r) {
                 // just make sure they are all not empty
-                $this->assertNotEmpty($r['value']);
+                $this->assertNotEmpty($r['value'],"Value is empty, looks like: ".var_export($r,true));
                 // make sure that the test field has our value in it
                 if($r['name'] == "test_custom_c") {
-                    $this->assertEquals("Custom Field", $r['value']);
+                    $this->assertEquals("Custom Field", $r['value'],"Custom field does not have our value in it");
                 }
             }
         } // if
